core: Use resource description to initialise OP-TEE
OP-TEE describes its resources to TF-A in a device tree as the
FVP_TOS_FW_CONFIG. The SPM consumes this description and translates it
into a SPCI initalisation message. This message is populated in the
OP-TEE secure RX buffer and its address is passed in 'x1'.

This patch adds support in OP-TEE to use this message to initialise its
translation tables. It also adds SPCI hooks to record the physical and
virtual addresses of the RX/TX buffers.

Original-author: Achin Gupta <achin.gupta@arm.com>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>