SPMC: add support to register secondary CPU entrypoints using PSCI_CPU_ON
This patch adds support to use the PSCI_CPU_ON function to register the entry
point for each OP-TEE context on a secondary CPU. This function is invoked on
the boot CPU during initialisation. When the physical CPU is turned on by the
Normal world, the SPMD in EL3 arranges for the entry point to be invoked to
perform OP-TEE initialisation.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>