core: enable SPCI with SPM Core at S-EL2
This is a port of a pre-exisiting patch to support the case where
the SPM Core is implementation at S-EL2, that is, in a secure hypervisor.
This configuration is also know as "S-EL2 SPMC" in the FFA specification.

Compile with CFG_CORE_SEL2_SPMC=y

Note that this is an experimental feature, ABIs etc may have
incompatible changes.

Sqaushed in:
	core: Update cookie to uint64 in S-EL2 code

	As part of the update to FF-A 1.0 the globally
	unique handle to identify a shared memory object is now 64
	bits wide instead of 32. This commit updates the
	remaining usage of 32 bit values.

	core: Remove page count from S-EL2 code

	The page count field has been dropped from upstream OP-TEE to
	accommodate using a 64 bit handle in the same size message param struct,
	therefore update our implementation to support this.

	core: Update remaining FF-A structures to EAC

	Update relinquish data structures and ABI invocations to the FFA
	EAC spec