core: ibmtss: include netinet/in.h for TPM_SKIBOOT
The TPM_SKIBOOT define is set when compiling for the SKIBOOT firmware,
but this is also very close to what OP-TEE needs so we are reusing the
define.

For the cases where the include directive for netinet/in.h is guarded
with TPM_POSIX add "|| defined(TPM_SKIBOOT)" to provide the endian
handling macros.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>